+++ inc/libavz.h
@@ -1,7 +1,7 @@
 #ifndef __LIBAVZ_H__
 #define __LIBAVZ_H__
 
-#define __AVZ_VERSION__ 250711
+#define __AVZ_VERSION__ 250823
 
 #include "avz_asm.h"
 #include "avz_card.h"